t0613: stop setting default initial branch
As the tests are all run in separate repositories, set the branch name
to "master" when creating the repository for the tests where the result
depends on the branch name. This reduceds the number of tests that
depend on the default branch name being "master" and removes the last
instance of a test file using "GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME=
master".

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>